requirements.txt
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.9" and python_version < "4"
dicom-validator==0.3.4 ; python_version >= "3.9" and python_version < "4.0"
dotty-dict==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gear-toolkit==0.5.9 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gears==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-sdk==16.16.7 ; python_version >= "3.9" and python_version < "4.0"
fw-core-client==1.1.2 ; python_version >= "3.9" and python_version < "4.0"
fw-file==1.4.2 ; python_version >= "3.9" and python_version < "4.0"
fw-http-client==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
fw-meta==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
fw-utils==4.2.7 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
jsonschema==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
memoization==0.4.0 ; python_version >= "3.9" and python_version < "4"
natsort==8.3.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.9" and python_version < "4.0"
pathvalidate==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.9" and python_version < "4.0"
pydicom==2.3.1 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==0.10.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rfc3987==1.3.8 ; python_version >= "3.9" and python_version < "4.0"
setuptools==67.6.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.9" and python_version < "4.0"
tzlocal==4.2 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4.0"
